fixed wrong js loading for eventify and publicEvents event methods #134
There was a bug with setting the eventify variable on `renderer.js`.
Instead of ngraph.events, the variable is set from the `ngraph.graph`
After this fix, publicEvents object has on and off methods to register
events.

event registration methods for the renderer erroneously uses
`publicEvents.addEventListener` and `publicEvents.removeEventListener`
method. changed the method name to `on` and `off`, respectively.