Add direct link to ngraph repo in readme
As a new user, I was confused about the `ngraph` family until I ran across the main repo page. It did not show up on the first page of results when I viewed the https://www.npmjs.com/search?q=ngraph link. I failed to understand that the core layout algorithms are decoupled from VivaGraphJS.

This commit adds a link to the main repo in the readme and tries to specify that it contains the layout algorithms.