Proposal: provide a possibility of passing custom zoom factor to renderer (not graphics)
Before this proposal one could pass custom factor to 'graphics.scale' BUT calling this method directly causes:
 - no redendering;
 - no raising 'scale' event by renderer;
 - no changing 'transform.scale'.

Of course these actions could be done later in third party code, but in this case it should just repeat the existing 'renderer.scale' method. It's better to allow developers not to do so.